Handle a case where an issue with no nodes breaks
Pa11y was exiting in the rare scenario when an aXe issue has no nodes.
This meant that we were attempting to select an element using an empty
string as a selector, which is invalid.

Now if there are no nodes for an issue, we set the issue element to
`null`, which is handled correctly by Pa11y.